Refactor under skjema
