+++ web/src/frontend/src/skjema/faktum/FaktumInput.less
