+++ web/src/frontend/src/skjema/faktum/faktumSkjemagruppe.less
