+++ web/src/frontend/src/skjema/reducer.ts
