+++ web/src/frontend/src/skjema/types.ts
@@ -1,4 +1,4 @@
-export * from "../../redux/types";
+export * from "../redux/types";
 
 export enum ActionTypeKeys {
 	SET_FAKTUM_VERDI = "SET_FAKTUM_VERDI",
